+++ core/options.go
@@ -353,6 +353,7 @@ type PipelineOptions struct {
 	ShouldArtifacts   bool
 	ShouldRemove      bool
 	SourceDir         string
+	IgnoreFile        string
 
 	AttachOnError  bool
 	DirectMount    bool
@@ -362,6 +363,12 @@ type PipelineOptions struct {
 	EnableVolumes  bool
 	WerckerYml     string
 	Checkpoint     string
+
+	DefaultsUsed PipelineDefaultsUsed
+}
+
+type PipelineDefaultsUsed struct {
+	IgnoreFile bool
 }
 
 func guessApplicationID(c util.Settings, e *util.Environment, name string) string {
@@ -551,6 +558,7 @@ func NewPipelineOptions(c util.Settings, e *util.Environment) (*PipelineOptions,
 	shouldRemove, _ := c.Bool("no-remove")
 	shouldRemove = !shouldRemove
 	sourceDir, _ := c.String("source-dir")
+	ignoreFile, ignoreFileSet := c.String("ignore-file")
 
 	attachOnError, _ := c.Bool("attach-on-error")
 	directMount, _ := c.Bool("direct-mount")
@@ -561,6 +569,10 @@ func NewPipelineOptions(c util.Settings, e *util.Environment) (*PipelineOptions,
 	enableVolumes, _ := c.Bool("enable-volumes")
 	werckerYml, _ := c.String("wercker-yml")
 	checkpoint, _ := c.String("checkpoint")
+
+	defaultsUsed := PipelineDefaultsUsed {
+		IgnoreFile: !ignoreFileSet,
+	}
 
 	return &PipelineOptions{
 		GlobalOptions: globalOpts,
@@ -602,6 +614,7 @@ func NewPipelineOptions(c util.Settings, e *util.Environment) (*PipelineOptions,
 		ShouldArtifacts:   shouldArtifacts,
 		ShouldRemove:      shouldRemove,
 		SourceDir:         sourceDir,
+		IgnoreFile:        ignoreFile,
 
 		AttachOnError:  attachOnError,
 		DirectMount:    directMount,
@@ -612,6 +625,8 @@ func NewPipelineOptions(c util.Settings, e *util.Environment) (*PipelineOptions,
 		EnableVolumes: enableVolumes,
 		WerckerYml:    werckerYml,
 		Checkpoint:    checkpoint,
+
+		DefaultsUsed: defaultsUsed,
 	}, nil
 }
 
@@ -679,6 +694,17 @@ func (o *PipelineOptions) ProjectDownloadPath() string {
 // StepPath returns the path where downloaded steps live
 func (o *PipelineOptions) StepPath() string {
 	return path.Join(o.WorkingDir, "steps")
+}
+
+// IgnoreFilePath return the absolute path of the ignore file
+func (o *PipelineOptions) IgnoreFilePath() string {
+	expandedIgnoreFile := util.ExpandHomePath(o.IgnoreFile, o.HostEnv.Get("HOME"))
+
+	if filepath.IsAbs(expandedIgnoreFile) {
+	  return expandedIgnoreFile
+	} else {
+	  return path.Join(o.ProjectPath, o.IgnoreFile)
+	}
 }
 
 // Options per Command